# apps/api/docs/swagger.json
{
    "schemes": ["http", "https"],
    "swagger": "2.0",
    "info": {
        "description": "Outdoor Equipment Rental Marketplace API",
        "title": "Rent Outdoor API",
        "contact": {
            "name": "Rent Outdoor API Support",
            "email": "support@rentoutdoor.id"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/stores": {
            "get": {
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "List stores",
                "description": "List active stores with optional filters.",
                "parameters": [
                    { "in": "query", "name": "city", "type": "string", "description": "Filter by city" },
                    { "in": "query", "name": "province", "type": "string", "description": "Filter by province" },
                    { "in": "query", "name": "search", "type": "string", "description": "Search by name or city" },
                    { "in": "query", "name": "page", "type": "integer", "default": 1, "description": "Page number" },
                    { "in": "query", "name": "per_page", "type": "integer", "default": 20, "description": "Items per page" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } }
                }
            },
            "post": {
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Create a new store",
                "description": "Creates a new store for the authenticated owner. One store per owner.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/request.CreateStoreRequest" } }
                ],
                "responses": {
                    "201": { "description": "Created", "schema": { "$ref": "#/definitions/response.Response" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "401": { "description": "Unauthorized", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "409": { "description": "Conflict", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/me": {
            "get": {
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Get my store",
                "description": "Get the authenticated owner's store.",
                "security": [{ "BearerAuth": [] }],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/{id}": {
            "get": {
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Get store by ID",
                "description": "Get detailed store information including photos and operating hours.",
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            },
            "put": {
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Update store",
                "description": "Update store information. Only the store owner can update.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/request.UpdateStoreRequest" } }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/slug/{slug}": {
            "get": {
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Get store by slug",
                "description": "Get store by its URL-friendly slug.",
                "parameters": [
                    { "in": "path", "name": "slug", "type": "string", "required": true, "description": "Store slug" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/{id}/photos": {
            "post": {
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Store Photos"],
                "summary": "Add store photo",
                "description": "Upload a photo for the store. Only the store owner can add photos.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/request.AddStorePhotoRequest" } }
                ],
                "responses": {
                    "201": { "description": "Created", "schema": { "$ref": "#/definitions/response.Response" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/{id}/photos/{photoId}": {
            "delete": {
                "produces": ["application/json"],
                "tags": ["Store Photos"],
                "summary": "Remove store photo",
                "description": "Remove a photo from the store. Only the store owner can remove.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "path", "name": "photoId", "type": "string", "required": true, "description": "Photo ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/{id}/photos/{photoId}/primary": {
            "patch": {
                "produces": ["application/json"],
                "tags": ["Store Photos"],
                "summary": "Set primary store photo",
                "description": "Set a photo as the primary/cover photo for the store.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "path", "name": "photoId", "type": "string", "required": true, "description": "Photo ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/stores/{id}/operating-hours": {
            "get": {
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Get store operating hours",
                "description": "Get the operating hours for a store.",
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } }
                }
            },
            "put": {
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Store"],
                "summary": "Set store operating hours",
                "description": "Set operating hours for each day of the week. Replaces existing hours.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/request.SetOperatingHoursRequest" } }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "403": { "description": "Forbidden", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "422": { "description": "Unprocessable Entity", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/admin/stores/{id}/approve": {
            "patch": {
                "produces": ["application/json"],
                "tags": ["Store Admin"],
                "summary": "Approve a store (Admin)",
                "description": "Approve a pending store and set it to active.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/admin/stores/{id}/suspend": {
            "patch": {
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Store Admin"],
                "summary": "Suspend a store (Admin)",
                "description": "Suspend an active store with a reason.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" },
                    { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/request.SuspendStoreRequest" } }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
        "/admin/stores/{id}/reactivate": {
            "patch": {
                "produces": ["application/json"],
                "tags": ["Store Admin"],
                "summary": "Reactivate a suspended store (Admin)",
                "description": "Reactivate a previously suspended store.",
                "security": [{ "BearerAuth": [] }],
                "parameters": [
                    { "in": "path", "name": "id", "type": "string", "required": true, "description": "Store ID" }
                ],
                "responses": {
                    "200": { "description": "OK", "schema": { "$ref": "#/definitions/response.Response" } },
                    "400": { "description": "Bad Request", "schema": { "$ref": "#/definitions/response.ErrorResponse" } },
                    "404": { "description": "Not Found", "schema": { "$ref": "#/definitions/response.ErrorResponse" } }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateStoreRequest": {
            "type": "object",
            "required": ["name", "phone", "email", "address", "city", "province"],
            "properties": {
                "name": { "type": "string", "example": "Outdoor Gear Bandung" },
                "description": { "type": "string", "example": "Premium outdoor equipment rental" },
                "phone": { "type": "string", "example": "08123456789" },
                "email": { "type": "string", "example": "store@example.com" },
                "address": { "type": "string", "example": "Jl. Raya Dago No. 123" },
                "city": { "type": "string", "example": "Bandung" },
                "province": { "type": "string", "example": "Jawa Barat" },
                "postal_code": { "type": "string", "example": "40135" },
                "latitude": { "type": "number", "example": -6.8845 },
                "longitude": { "type": "number", "example": 107.6145 }
            }
        },
        "request.UpdateStoreRequest": {
            "type": "object",
            "properties": {
                "name": { "type": "string", "example": "Updated Store Name" },
                "description": { "type": "string" },
                "phone": { "type": "string" },
                "email": { "type": "string" },
                "address": { "type": "string" },
                "city": { "type": "string" },
                "province": { "type": "string" },
                "postal_code": { "type": "string" },
                "latitude": { "type": "number" },
                "longitude": { "type": "number" },
                "logo_url": { "type": "string" },
                "banner_url": { "type": "string" }
            }
        },
        "request.AddStorePhotoRequest": {
            "type": "object",
            "required": ["photo_url"],
            "properties": {
                "photo_url": { "type": "string", "example": "https://cdn.example.com/photo.jpg" },
                "caption": { "type": "string", "example": "Storefront view" },
                "sort_order": { "type": "integer", "example": 1 },
                "is_primary": { "type": "boolean", "example": false }
            }
        },
        "request.SetOperatingHoursRequest": {
            "type": "object",
            "required": ["hours"],
            "properties": {
                "hours": {
                    "type": "array",
                    "items": { "$ref": "#/definitions/request.OperatingHourItem" }
                }
            }
        },
        "request.OperatingHourItem": {
            "type": "object",
            "required": ["day_of_week", "open_time", "close_time"],
            "properties": {
                "day_of_week": { "type": "integer", "minimum": 1, "maximum": 7, "example": 1 },
                "open_time": { "type": "string", "example": "08:00" },
                "close_time": { "type": "string", "example": "17:00" },
                "is_closed": { "type": "boolean", "example": false }
            }
        },
        "request.SuspendStoreRequest": {
            "type": "object",
            "required": ["reason"],
            "properties": {
                "reason": { "type": "string", "example": "Violation of terms of service" }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean", "example": true },
                "message": { "type": "string", "example": "operation successful" },
                "data": { "type": "object" },
                "meta": { "type": "object" }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "success": { "type": "boolean", "example": false },
                "message": { "type": "string", "example": "error message" },
                "errors": { "type": "object" }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}